Codebase update + changed file/class names + created separate chart setup file for convenience
