+++ Alames/__init__.py
@@ -2,7 +2,7 @@ import sys, os
 from PyQt5.QtWidgets import QApplication
 from PyQt5.QtGui import QFontDatabase, QFont
 
-from Alames import main
+from Alames.window import Window
 
 def launch():
     app = QApplication(sys.argv)
@@ -10,5 +10,5 @@ def launch():
     # app.setWindowIcon(QtGui.QIcon("icons/main.png"))
     QFontDatabase.addApplicationFont(os.path.join(os.path.join(os.path.dirname(__file__), "fonts"), "Gidole-Regular.ttf"))
     app.setFont(QFont("Gidole"))
-    ex = main.Window(app)
+    Window(app)
     sys.exit(app.exec_())
